# microservices/core-rest-service/docs/openapi.yaml
openapi: 3.0.3

info:
  title: Core REST Service
  description: REST API for the Factory Manager distributed monitoring system.
  version: 0.1.0

servers:
  - url: http://localhost:3000
    description: Local development

tags:
  - name: Auth
    description: Authentication
  - name: Users
    description: User management

paths:
  /api/auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: Log in and obtain a JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/auth/logout:
    post:
      tags: [Auth]
      operationId: logout
      summary: Log out (stateless — client discards the token)
      responses:
        '204':
          description: Logged out

  /api/users:
    post:
      tags: [Users]
      operationId: createUser
      summary: Create a user
      description: Requires an authenticated admin user.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'

    get:
      tags: [Users]
      operationId: listUsers
      summary: List users
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/users/{id}:
    parameters:
      - $ref: '#/components/parameters/UserId'

    get:
      tags: [Users]
      operationId: getUserById
      summary: Get a user by ID
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    patch:
      tags: [Users]
      operationId: updateUserById
      summary: Update a user by ID
      description: Requires an authenticated admin user.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      tags: [Users]
      operationId: deleteUserById
      summary: Delete a user by ID
      description: Requires an authenticated admin user.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    UserId:
      name: id
      in: path
      required: true
      description: 24-character hexadecimal MongoDB ObjectId
      schema:
        type: string
        pattern: '^[0-9a-f]{24}$'
        example: '507f1f77bcf86cd799439011'

  schemas:
    UserName:
      type: object
      required: [first, last]
      properties:
        first:
          type: string
          example: Mario
        last:
          type: string
          example: Rossi

    User:
      type: object
      required: [id, name, email, role, isActive, lastLoginAt, createdAt, updatedAt, fullName]
      properties:
        id:
          type: string
          example: '507f1f77bcf86cd799439011'
        name:
          $ref: '#/components/schemas/UserName'
        email:
          type: string
          format: email
          example: mario.rossi@fm.com
        role:
          type: string
          enum: [admin, operator]
          example: operator
        isActive:
          type: boolean
          example: true
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
          example: null
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        fullName:
          type: string
          example: Mario Rossi

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: mario.rossi@fm.com
        password:
          type: string
          format: password
          example: Password123!

    LoginResponse:
      type: object
      required: [accessToken, tokenType, expiresIn, user]
      properties:
        accessToken:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.e30.signature
        tokenType:
          type: string
          example: Bearer
        expiresIn:
          type: string
          example: 1h
        user:
          $ref: '#/components/schemas/User'

    CreateUserRequest:
      type: object
      required: [name, email, password]
      properties:
        name:
          $ref: '#/components/schemas/UserName'
        email:
          type: string
          format: email
          example: mario.rossi@fm.com
        password:
          type: string
          format: password
          minLength: 8
          pattern: '^(?=.*[A-Za-z])(?=.*\d)(?=.*[@$!%*#?&])[A-Za-z\d@$!%*#?&]{8,}$'
          description: >
            At least 8 characters containing one letter, one number, and one
            special character (@$!%*#?&).
          example: Password123!
        role:
          type: string
          enum: [admin, operator]
          example: operator

    UpdateUserRequest:
      type: object
      minProperties: 1
      properties:
        name:
          type: object
          minProperties: 1
          properties:
            first:
              type: string
              example: Mario
            last:
              type: string
              example: Rossi
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 8
          pattern: '^(?=.*[A-Za-z])(?=.*\d)(?=.*[@$!%*#?&])[A-Za-z\d@$!%*#?&]{8,}$'
          description: >
            At least 8 characters containing one letter, one number, and one
            special character (@$!%*#?&).
        role:
          type: string
          enum: [admin, operator]
        isActive:
          type: boolean

    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            [
              VALIDATION_ERROR,
              NOT_FOUND,
              CONFLICT,
              UNAUTHORIZED,
              FORBIDDEN,
              INTERNAL_ERROR
            ]
          example: NOT_FOUND
        message:
          type: string
          example: User not found
        details:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field:
                type: string
                example: email
              message:
                type: string
                example: '"email" must be a valid email'

  responses:
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: VALIDATION_ERROR
            message: '"email" must be a valid email'

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: NOT_FOUND
            message: User not found

    Conflict:
      description: Resource already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: CONFLICT
            message: A resource with this value already exists

    Unauthorized:
      description: Authentication required or credentials invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: UNAUTHORIZED
            message: Authentication required

    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: FORBIDDEN
            message: Insufficient permissions

    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: INTERNAL_ERROR
            message: Internal server error